+++ NEWS.md
@@ -1,3 +1,10 @@
+# sbo 0.1.2
+
+* Added `preproc` argument to `get_kgram_freqs()` and `get_word_freqs()`, for 
+custom training corpus preprocessing.
+* The `dict` argument of `get_kgram_freqs()` now also accepts numeric values,
+allowing to build a dictionary directly from the training corpus.
+
 # sbo 0.1.1
 
 * Added `predict` method for `kgram_freqs` class.